chore: remove comment & move file
